pyproject.toml
[project]
name = "dj-apis-allauth"
version = "1.0.0"
description = "Secure API Endpoints Authentication with Django REST Framework"
authors = [
    {name = "Prince Clumson-Eklu",email = "ameto888@protonmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
homepage = "https://github.com/pclumson/dj-apis-allauth"
repository = "https://github.com/pclumson/dj-apis-allauth"
documentation = "https://dj-apis-allauth.readthedocs.io"
keywords = ["pypi", "poetry", "packaging"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Framework :: Django :: 5.2",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "django-allauth (>=65.11.2,<66.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
